Refactor code and split into many files
Closes #113
